util/requirements.txt
python-dateutil
psycopg2
psycopg2-binary
urllib3
six
suds-py3
